Don't rely on "characters" when ranging over string
Should probably read https://blog.golang.org/strings again.

"To answer the question posed at the beginning: Strings are built from
bytes so indexing them yields bytes, not characters. A string might
not even hold characters. In fact, the definition of "character" is
ambiguous and it would be a mistake to try to resolve the ambiguity by
defining that strings are made of characters."

Thanks, Rob Pike.